The music now has its own controller.
<organisation
